--- src/App.jsx
@@ -1,15 +1,0 @@
-import Landing from "./components/Landing/Landing";
-import Navbar from "./components/Navbar/Navbar";
-import Footer from "./components/Footer/Footer";
-
-function App() {
-	return (
-		<div>
-			<Navbar />
-			<Landing />
-			<Footer />
-		</div>
-	);
-}
-
-export default App;
