--- src/Pages/Home/Home.jsx
@@ -1,10 +1,0 @@
-import { Navbar, Footer } from "../../components/index";
-const Home = () => {
-	return (
-		<>
-			<Navbar />
-			<Footer />
-		</>
-	);
-};
-export default Home;
