+++ src/components/Landing/Landing.jsx
@@ -8,9 +8,17 @@ const Landing = () => {
 			<div>
 				<img src="/Variant3.svg" alt="" className={styles.biome_text} />
 			</div>
-			<div className={styles.arrow}>
-				<Lottie animationData={arrowAnimation} />
+			<div className={styles.superflex}>
+				<img src="/Subtract.png" alt="" className={styles.tag} />
+				<img src="/nitshack.png" alt="" className={styles.nitshack} />
+				<img src="/dates.png" alt="" className={styles.dates} />
+				<div className={styles.arrow}>
+					<a href="#about">
+						<Lottie animationData={arrowAnimation} />
+					</a>
+				</div>
 			</div>
+			<div className={styles.empty}></div>
 		</div>
 	);
 };
