+++ src/components/Landing/landing.module.scss
@@ -1,28 +1,66 @@
 .landing_main {
-  max-width: 100%;
-  height: 100vh;
-  background-image: url(/landing_main.svg);
+  width: 100%;
+  height: 120vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-image: linear-gradient(to top, black, transparent),
+    url(landing_main.svg);
   background-size: cover;
-  filter: drop-shadow();
+  background-position: center bottom;
   background-repeat: no-repeat;
-  overflow-x: hidden;
+  filter: saturate(1.6);
+  overflow: hidden;
 
-  .tag {
-    width: 50rem;
-    height: 70rem;
-  }
-  .arrow {
-    position: absolute;
-    bottom: 2rem;
-    left: 47%;
-    transform: rotateZ(270deg);
-    width: 8rem;
-    cursor: pointer;
-  }
   .biome_text {
-    position: relative;
-    left: -0.5rem;
-    bottom: 0.5rem;
-    height: 12rem;
+    position: absolute;
+    left: -2.7rem;
+    top: -1.2rem;
+    width: 24rem;
+  }
+
+  .superflex {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    .tag {
+      width: 90rem;
+      position: absolute;
+      top: -1rem;
+      @media screen and (max-width: 440px) {
+        width: 55rem;
+        height: 35rem;
+      }
+      @media screen and (max-width: 1024px) {
+        top: 9rem;
+      }
+    }
+    .nitshack {
+      margin-top: 30rem;
+      width: 30rem;
+      z-index: 1;
+      @media screen and (max-width: 440px) {
+        width: 20rem;
+        margin-top: 20rem;
+      }
+      @media screen and (max-width: 390px) {
+        width: 20rem;
+        margin-top: 30rem;
+      }
+    }
+    .dates {
+      margin-top: -7rem;
+      @media screen and (max-width: 440px) {
+        width: 20rem;
+        margin-top: -4rem;
+      }
+    }
+    .arrow {
+      width: 8rem;
+      transform: rotateZ(-90deg);
+      cursor: pointer;
+      z-index: 1;
+    }
   }
 }
