+++ src/components/Navbar/Navbar.jsx
@@ -1,4 +1,6 @@
+import styles from "./Navbar.module.scss";
+
 const Navbar = () => {
-	return <div>Navbar</div>;
+	return <div className={styles.nav}>Navbar</div>;
 };
 export default Navbar;
