+++ src/components/Navbar/Navbar.module.scss
@@ -1,0 +1,8 @@
+.nav {
+  width: 100%;
+  height: 5rem;
+  position: fixed;
+  z-index: 1;
+  background-image: linear-gradient(black, transparent);
+  color: white;
+}
