--- src/components/Preview/preview.module.scss
@@ -1,74 +1,0 @@
-.container {
-  max-width: 100%;
-  height: 100vh;
-  background-image: url(/space.png);
-  background-size: contain;
-  padding: 2.3rem;
-
-  .glow_container {
-    border: 0.4rem solid #d03cab;
-    box-shadow: 1px 2px 3px 0.2rem #ce72b7;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    .text {
-      display: flex;
-      width: 100%;
-      padding: 1.5rem;
-      justify-content: space-between;
-
-      @media screen and (max-width: 1300px) {
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        gap: 0.5rem;
-      }
-      .bit,
-      .transfer {
-        height: 28px;
-        width: 30%;
-
-        @media screen and (max-width: 1024px) {
-          width: 80%;
-        }
-
-        @media screen and (max-width: 767px) {
-          width: 100%;
-          height: 16px;
-        }
-      }
-    }
-
-    .insignia {
-      img {
-        position: relative;
-        height: 25rem;
-        @media screen and (max-width: 1024px) {
-          top: 11rem;
-          height: 22rem;
-        }
-        @media screen and (max-width: 480px) {
-          top: 15rem;
-          height: 10rem;
-        }
-        @media screen and (max-width: 380px) {
-          top: 8rem;
-          height: 10rem;
-        }
-      }
-    }
-    .upgrading {
-      background-color: black;
-      padding: 0rem 0.7rem;
-      position: absolute;
-      width: 30rem;
-      bottom: 1rem;
-      @media screen and (max-width: 480px) {
-        width: 15rem;
-        bottom: 1.5rem;
-      }
-    }
-  }
-}
